Fix packages not using macros from buck2/whatsapp/erlang.bzl
Summary:
# Context
We are adding a new linter that will detect when a package define `erlang_app` or `erlang_test` targets that didn't come from the macros in `buck2/whatsapp/erlang.bzl`.

# Problem
Running the linter on all packages shows packages with failures

# This diff
Fix the packages to import the required macros. Notice that some of the packages were unit-tests for linters, we are fix them or otherwise their golden tests would fail as we would now emit also the new linter failures

Differential Revision: D56254472

fbshipit-source-id: cce56604720ab7bd2e3c6d87323c76718fd647eb